requirements.txt
psutil==6.0.0
paramiko==3.4.0
https.server==1.2.0
pystray==0.19.5
requests==2.32.3